+++ mblt_model_zoo/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.5.2"
+__version__ = "2.5.3"
 from . import utils, vision
 
 __all__ = ["utils", "vision"]